requirements.txt
# ─── Flask core ───────────────────────────────────────────────────────────────
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Migrate==4.0.7

# ─── Database ─────────────────────────────────────────────────────────────────
PyMySQL==1.1.1
SQLAlchemy==2.0.30
cryptography==42.0.8

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
ollama==0.2.1
requests==2.32.3
httpx==0.27.0

# ─── Embeddings & NLP ─────────────────────────────────────────────────────────
transformers==4.41.2
torch==2.3.1
sentence-transformers==3.0.1
tokenizers==0.19.1

# ─── FAISS ────────────────────────────────────────────────────────────────────
faiss-cpu==1.8.0
numpy==1.26.4

# ─── Data processing ──────────────────────────────────────────────────────────
pandas==2.2.2
scikit-learn==1.5.0

# ─── Auth & Security ──────────────────────────────────────────────────────────
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1

# ─── Validation ───────────────────────────────────────────────────────────────
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
email-validator==2.2.0

# ─── PDF Export ───────────────────────────────────────────────────────────────
reportlab==4.2.2
fpdf2==2.7.9

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
Werkzeug==3.0.3
gunicorn==22.0.0
celery==5.4.0
redis==5.0.6

# ─── Monitoring & Logging ─────────────────────────────────────────────────────
structlog==24.2.0
prometheus-flask-exporter==0.23.1

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-flask==1.3.0
pytest-cov==5.0.0
factory-boy==3.3.0
